(refactor): regroup crawling module functionality
CTINFRA-3025
