+++ src/resources/Connectivity/index.ts
@@ -1,3 +1,2 @@
 export * from './Connectivity';
-export * from './CrawlingModules';
 export * from './ConnectivityInterface';
