+++ src/resources/CrawlingModule/CrawlingModuleInterfaces.ts
@@ -30,3 +30,37 @@ export interface UpdateStatus {
     updateAvailable: boolean;
     updateCategory: UpdateStatusCategory;
 }
+
+export enum CrawlingModuleLogRequestLogType {
+    MAESTRO = 'MAESTRO',
+    WORKER = 'WORKER',
+}
+
+export enum CrawlingModuleLogRequestState {
+    CREATED = 'CREATED',
+    ERROR = 'ERROR',
+    RUNNING = 'RUNNING',
+    SUCCESSFUL = 'SUCCESSFUL',
+    TIMEOUT = 'TIMEOUT',
+}
+
+export interface CrawlingModuleLogRequestModel {
+    createdDate: number;
+    id: string;
+    instanceId: string;
+    logType: CrawlingModuleLogRequestLogType;
+    operationId: string;
+    state: CrawlingModuleLogRequestState;
+}
+
+export interface CreateCrawlingModuleLogRequestModel {
+    instanceId: string;
+    logType: CrawlingModuleLogRequestLogType;
+    operationId: string;
+}
+
+export interface CrawlingModuleLogRequestDownloadModel {
+    info: string;
+    state: CrawlingModuleLogRequestState;
+    url: string;
+}
